fix(upload): return type on is now a string (#976)
* fix(upload): return type on POST is now string

A POST to a webserver can not always be expected to be a JSON response.

Success is now determined by the HTTP return code.

Upon success the body content is returned as a string.

* feat: add content-length on POST

Not all embedded devices are acceptable to receiving unspecified amounts
of data. Appending the content-length up front helps this devices
succeed.

* fix: return values unified

The return values was not used.

On POST the HTTP error code is returned as an enum.

* fix: upload, return value as string

* Update plugins/upload/src/lib.rs

Co-authored-by: Lucas Fernandes Nogueira <lucas@tauri.app>

* Update plugins/upload/src/lib.rs

Co-authored-by: Lucas Fernandes Nogueira <lucas@tauri.app>

* fix: added covector changelog file

---------

Co-authored-by: Lucas Fernandes Nogueira <lucas@tauri.app>

Committed via a GitHub action: https://github.com/tauri-apps/plugins-workspace/actions/runs/8689831151

Co-authored-by: FabianLars <FabianLars@users.noreply.github.com>